Moved around lib files toi allow splitting files in memory
