+++ lib/audio/audio.js
