+++ lib/controller/controller.js
@@ -1,5 +1,5 @@
 import Promise from 'promise-polyfill';
-import { WASMBOY_CONTROLLER_STATE } from './wasmboyControllerState';
+import { WASMBOY_CONTROLLER_STATE } from './controllerState';
 
 // Helpers for accessing gamepad
 // Similar to: https://github.com/torch2424/picoDeploy/blob/master/src/assets/3pLibs/pico8gamepad/pico8gamepad.js
