+++ lib/wasmboy.js
@@ -1,10 +1,10 @@
 import fetch from 'unfetch'
 import Promise from 'promise-polyfill';
 import wasmModule from '../dist/wasm/index.untouched.wasm';
-import { WasmBoyGraphics } from './wasmboyGraphics';
-import { WasmBoyAudio } from './wasmboyAudio';
-import { WasmBoyController } from './wasmboyController';
-import { WasmBoyMemory } from './wasmboyMemory';
+import { WasmBoyGraphics } from './graphics/graphics';
+import { WasmBoyAudio } from './audio/audio';
+import { WasmBoyController } from './controller/controller';
+import { WasmBoyMemory } from './memory/memory';
 
 // Polyfill for raf for testing in node, defaults to builtin window.requestAnimationFrame
 const raf = require('raf');
@@ -294,7 +294,7 @@ class WasmBoyLib {
 }
 
 export const WasmBoy = new WasmBoyLib();
-export { WasmBoyAudio } from './wasmboyAudio';
-export { WasmBoyGraphics } from './wasmboyGraphics';
+export { WasmBoyAudio } from './audio/audio';
+export { WasmBoyGraphics } from './graphics/graphics';
 // TODO: Remove this, and consolidate public api in Wasmboy
-export { WasmBoyMemory } from './wasmboyMemory';
+export { WasmBoyMemory } from './memory/memory';
